\name{GSALightning-package}
\alias{GSALightning-package}
\alias{GSALightning}
\docType{package}
\title{
\packageTitle{GSALightning}
}
\description{
\packageDescription{GSALightning}
}
\details{

\packageIndices{GSALightning}

}
\author{
\packageAuthor{GSALightning}

Maintainer: \packageMaintainer{GSALightning}
}
\references{
BHW Chang and W Tian (2015). GSA-Lightning: Ultra Fast Permutation-based Gene Set Analysis. (Submitted)

B Efron and RJ Tibshirani (2007). "On testing the significance of sets of genes." The annals of applied statistics 1(1):107-129
}

\keyword{ package }
\seealso{
\code{\link{GSALight}}, \code{\link{permTestLight}}.
}
